More cosmetics, ctypes for image copy
